2025 day 10: Get something that almost works...
This works for the first two examples, but not the final example.
Finding a valid solution is all well and good, but we need to find an
_optimal_ solution.

Even if there is an optimal way to press the initial buttons for parity,
we need to check _all_ optimal ways to press those buttons.